+++ db/migrate/20160115214650_create_developer_scores.rb
@@ -1,0 +1,5 @@
+class CreateDeveloperScores < ActiveRecord::Migration
+  def change
+    create_view :developer_scores
+  end
+end