+++ db/schema.rb
@@ -65,4 +65,58 @@ ActiveRecord::Schema.define(version: 20160205153837) do
 
   add_foreign_key "posts", "channels"
   add_foreign_key "posts", "developers"
+
+  create_view :hot_posts,  sql_definition: <<-SQL
+      WITH posts_with_age AS (
+           SELECT posts.id,
+              posts.developer_id,
+              posts.body,
+              posts.created_at,
+              posts.updated_at,
+              posts.channel_id,
+              posts.title,
+              posts.slug,
+              posts.likes,
+              posts.tweeted,
+              posts.published_at,
+              GREATEST((date_part('epoch'::text, (now() - posts.published_at)) / (3600)::double precision), (0.1)::double precision) AS hour_age
+             FROM posts
+            WHERE (posts.published_at IS NOT NULL)
+          )
+   SELECT ((posts_with_age.likes)::double precision / (posts_with_age.hour_age ^ (0.8)::double precision)) AS score,
+      posts_with_age.id,
+      posts_with_age.developer_id,
+      posts_with_age.body,
+      posts_with_age.created_at,
+      posts_with_age.updated_at,
+      posts_with_age.channel_id,
+      posts_with_age.title,
+      posts_with_age.slug,
+      posts_with_age.likes,
+      posts_with_age.tweeted,
+      posts_with_age.published_at,
+      posts_with_age.hour_age
+     FROM posts_with_age
+    ORDER BY ((posts_with_age.likes)::double precision / (posts_with_age.hour_age ^ (0.8)::double precision)) DESC;
+  SQL
+
+  create_view :developer_scores,  sql_definition: <<-SQL
+      SELECT developers.id,
+      developers.username,
+      stats.posts,
+      stats.likes,
+      round(((stats.likes)::numeric / (stats.posts)::numeric), 2) AS avg_likes,
+      round(log((2)::numeric, ((((1022)::double precision * ((developer_scores.score - min(developer_scores.score) OVER ()) / (max(developer_scores.score) OVER () - min(developer_scores.score) OVER ()))) + (2)::double precision))::numeric), 1) AS hotness
+     FROM ((developers
+       JOIN ( SELECT hot_posts.developer_id AS id,
+              sum(hot_posts.score) AS score
+             FROM hot_posts
+            GROUP BY hot_posts.developer_id) developer_scores USING (id))
+       JOIN ( SELECT posts.developer_id AS id,
+              count(*) AS posts,
+              sum(posts.likes) AS likes
+             FROM posts
+            GROUP BY posts.developer_id) stats USING (id));
+  SQL
+
 end